+++ frontend/src/views/bpm/tasks/index.vue
@@ -3,7 +3,7 @@
     <el-card>
       <template #header>
         <div class="flex justify-between items-center">
-          <span>My Tasks (Element Plus)</span>
+          <span>我的待办</span>
           <el-button type="primary" @click="fetchTasks">Refresh</el-button>
         </div>
       </template>